feat:对接前端报警配置，增加.gitignore文件内容

+++ config/src/main/java/io/dynamic/threadpool/config/service/biz/ConfigServiceImpl.java
@@ -62,7 +62,7 @@ public class ConfigServiceImpl implements ConfigService {
                 .notifyConfigChange(new LocalDataChangeEvent(identify, ContentUtil.getGroupKey(configAllInfo)));
     }
 
-    private Integer addConfigInfo(ConfigAllInfo config) {
+    private Long addConfigInfo(ConfigAllInfo config) {
         config.setContent(ContentUtil.getPoolContent(config));
         config.setMd5(Md5Util.getTpContentMd5(config));
         try {
